requirements.txt
torchvision
h5py
numpy
matplotlib
transformers
pillow
tqdm
requests
torch
datasets